+++ system/kubernetes/apps/init.go
@@ -1,0 +1,13 @@
+package core
+
+import (
+	"github.com/viant/endly"
+	_ "github.com/viant/endly/system/kubernetes/apps/v1"
+)
+
+
+func init() {
+	_ = endly.Registry.Register(func() endly.Service {
+		return New()
+	})
+}