+++ system/kubernetes/apps/service_test.go
@@ -1,0 +1,11 @@
+package core
+
+import (
+	"github.com/stretchr/testify/assert"
+	"testing"
+)
+
+func TestNew(t *testing.T) {
+	service := New()
+	assert.NotNil(t, service)
+}