+++ system/kubernetes/apps/v1/adapter.go
@@ -1,0 +1,5 @@
+package v1
+
+import "github.com/viant/endly/system/kubernetes/registry"
+
+var register = registry.Register